requirements.txt
customtkinter
pillow
tkcalendar
ttkbootstrap
tkinterdnd2
